+++ android/app/src/main/java/com/emergetools/hackernews/features/bookmarks/BookmarksScreen.kt
@@ -1,47 +1,28 @@
 package com.emergetools.hackernews.features.bookmarks
 
-import androidx.compose.animation.core.Animatable
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawWithContent
-import androidx.compose.ui.draw.scale
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
+import com.emergetools.hackernews.features.bookmarks.components.BookmarksEducationCard
 import com.emergetools.hackernews.features.comments.CommentsDestinations
-import com.emergetools.hackernews.features.stories.ListSeparator
 import com.emergetools.hackernews.features.stories.StoriesDestinations
 import com.emergetools.hackernews.features.stories.StoryItem
-import com.emergetools.hackernews.features.stories.StoryRow
+import com.emergetools.hackernews.ui.components.ColumnSeparator
+import com.emergetools.hackernews.ui.components.StoryRow
 import com.emergetools.hackernews.ui.theme.HackerNewsTheme
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
 
 @Composable
 fun BookmarksScreen(
@@ -64,7 +45,9 @@ fun BookmarksScreen(
       style = MaterialTheme.typography.titleMedium
     )
     if (state.bookmarks.isEmpty()) {
-      BookmarksEmptyState()
+      Box(modifier = Modifier.fillMaxWidth().weight(1f), contentAlignment = Alignment.Center) {
+        BookmarksEducationCard()
+      }
     } else {
       LazyColumn {
         itemsIndexed(items = state.bookmarks, key = { _, item -> item.id }) { index, item ->
@@ -83,7 +66,7 @@ fun BookmarksScreen(
             }
           )
           if (index != state.bookmarks.lastIndex) {
-            ListSeparator(
+            ColumnSeparator(
               lineColor = MaterialTheme.colorScheme.surfaceContainerHighest,
             )
           }
@@ -119,6 +102,7 @@ fun BookmarksScreenPreview() {
             commentCount = 23,
             epochTimestamp = 100L,
             timeLabel = "2h ago",
+            bookmarked = true,
             url = ""
           ),
           StoryItem.Content(
@@ -129,6 +113,7 @@ fun BookmarksScreenPreview() {
             commentCount = 103,
             epochTimestamp = 100L,
             timeLabel = "7h ago",
+            bookmarked = true,
             url = ""
           ),
         )
@@ -153,117 +138,3 @@ fun BookmarksScreenEmptyPreview() {
   }
 }
 
-@Composable
-private fun ColumnScope.BookmarksEmptyState() {
-  Column(
-    modifier = Modifier
-      .fillMaxWidth()
-      .weight(1f),
-    verticalArrangement = Arrangement.spacedBy(
-      space = 8.dp,
-      alignment = Alignment.CenterVertically
-    ),
-    horizontalAlignment = Alignment.CenterHorizontally
-  ) {
-    val scale = remember { Animatable(0.8f) }
-    var bookmarked by remember { mutableStateOf(false) }
-    val nubRadius = remember { Animatable(30f) }
-    val nubTranslation = remember { Animatable(0.1f) }
-    val nubColor = MaterialTheme.colorScheme.onBackground.copy(0.2f)
-    LaunchedEffect(Unit) {
-      while (true) {
-        delay(3000)
-        nubTranslation.animateTo(0.3f)
-        launch {
-          scale.animateTo(0.7f)
-        }
-        launch {
-          nubRadius.animateTo(20f)
-        }
-        delay(700)
-        bookmarked = true
-        delay(500)
-        launch {
-          scale.animateTo(0.8f)
-        }
-        launch {
-          nubRadius.animateTo(30f)
-        }
-        delay(200)
-        nubTranslation.animateTo(0.1f)
-        delay(3000)
-        nubTranslation.animateTo(0.3f)
-        launch {
-          scale.animateTo(0.7f)
-        }
-        launch {
-          nubRadius.animateTo(20f)
-        }
-        delay(700)
-        bookmarked = false
-        delay(500)
-        launch {
-          scale.animateTo(0.8f)
-        }
-        launch {
-          nubRadius.animateTo(30f)
-        }
-        delay(200)
-        nubTranslation.animateTo(0.1f)
-      }
-    }
-    Text(
-      modifier = Modifier.wrapContentSize(),
-      text = "Long press a story to bookmark it...",
-      color = MaterialTheme.colorScheme.onBackground,
-      style = MaterialTheme.typography.labelMedium,
-      fontWeight = FontWeight.Medium
-    )
-    StoryRow(
-      modifier = Modifier
-        .drawWithContent {
-          drawContent()
-
-          val x = size.width * 0.8f
-          val y = size.height * nubTranslation.value
-          drawCircle(
-            color = nubColor,
-            center = Offset(x, y),
-            radius = nubRadius.value
-          )
-          drawCircle(
-            color = nubColor,
-            center = Offset(x, y),
-            radius = 30f
-          )
-        }
-        .scale(scale.value)
-        .border(
-          width = 1.dp,
-          color = MaterialTheme.colorScheme.onBackground,
-          shape = RoundedCornerShape(8.dp)
-        ),
-      item = StoryItem.Content(
-        id = 1L,
-        title = "Show HN: A new Android client",
-        author = "heyrikin",
-        score = 10,
-        commentCount = 45,
-        epochTimestamp = 100L,
-        timeLabel = "3h ago",
-        bookmarked = bookmarked,
-        url = ""
-      ),
-      onClick = {},
-      onBookmark = {},
-      onCommentClicked = {}
-    )
-    Text(
-      modifier = Modifier.wrapContentSize(),
-      text = "...and save it for later",
-      color = MaterialTheme.colorScheme.onBackground,
-      style = MaterialTheme.typography.labelMedium,
-      fontWeight = FontWeight.Medium
-    )
-  }
-}
