+++ android/app/src/main/java/com/emergetools/hackernews/features/comments/CommentsScreen.kt
@@ -1,12 +1,6 @@
 package com.emergetools.hackernews.features.comments
 
 import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.tween
 import androidx.compose.animation.slideInVertically
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -21,12 +15,9 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
@@ -36,7 +27,6 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -46,24 +36,17 @@ import androidx.compose.ui.geometry.center
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.PathEffect
 import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.emergetools.hackernews.R
-import com.emergetools.hackernews.features.stories.MetadataButton
-import com.emergetools.hackernews.features.stories.MetadataTag
-import com.emergetools.hackernews.parseAsHtml
-import com.emergetools.hackernews.ui.theme.HackerGreen
+import com.emergetools.hackernews.features.comments.components.CommentRow
+import com.emergetools.hackernews.features.comments.components.CommentsHeader
+import com.emergetools.hackernews.features.comments.components.PostCommentBump
 import com.emergetools.hackernews.ui.theme.HackerNewsTheme
-import com.emergetools.hackernews.ui.theme.HackerOrange
-import com.emergetools.hackernews.ui.theme.HackerPurple
 
 @Composable
 fun CommentsScreen(
@@ -82,7 +65,7 @@ fun CommentsScreen(
       verticalArrangement = Arrangement.spacedBy(4.dp)
     ) {
       item {
-        ItemHeader(
+        CommentsHeader(
           state = state.headerState,
           modifier = Modifier
             .fillMaxWidth()
@@ -237,510 +220,3 @@ private fun CommentsScreenLoadingPreview() {
   }
 }
 
-@Composable
-fun CommentRow(
-  modifier: Modifier = Modifier,
-  state: CommentState,
-  onToggleHide: (CommentState.Content) -> Unit,
-  onLikeTapped: (CommentState.Content) -> Unit
-) {
-  when (state) {
-    is CommentState.Content -> {
-      val startPadding = (state.level * 16).dp
-      Column(
-        modifier = modifier
-          .padding(start = startPadding)
-          .fillMaxWidth()
-          .wrapContentHeight()
-          .clip(RoundedCornerShape(8.dp))
-          .clickable { onToggleHide(state) }
-          .background(color = MaterialTheme.colorScheme.surface)
-          .padding(8.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-      ) {
-        Row(
-          modifier = Modifier.fillMaxWidth(),
-          horizontalArrangement = Arrangement.spacedBy(8.dp),
-          verticalAlignment = Alignment.CenterVertically
-        ) {
-          Text(
-            text = "@${state.author}",
-            style = MaterialTheme.typography.labelSmall,
-            color = MaterialTheme.colorScheme.onSurface,
-            fontWeight = FontWeight.Bold
-          )
-          MetadataTag(
-            label = state.timeLabel
-          ) {
-            Icon(
-              modifier = Modifier.size(12.dp),
-              painter = painterResource(R.drawable.ic_time_outline),
-              tint = MaterialTheme.colorScheme.onSurface,
-              contentDescription = "Time Posted"
-            )
-          }
-          Icon(
-            modifier = Modifier
-              .graphicsLayer {
-                rotationZ = if (state.hidden == HiddenStatus.Hidden) 180f else 0f
-              }
-              .size(12.dp),
-            painter = painterResource(R.drawable.ic_collapse),
-            tint = MaterialTheme.colorScheme.onSurface,
-            contentDescription = "Expand or Collapse"
-          )
-          Spacer(modifier = Modifier.weight(1f))
-          Box(
-            modifier = Modifier
-              .wrapContentSize()
-              .clip(CircleShape)
-              .clickable { onLikeTapped(state) }
-              .background(
-                color = if (state.upvoted) {
-                  HackerGreen.copy(alpha = 0.2f)
-                } else {
-                  MaterialTheme.colorScheme.surfaceContainerHighest
-                }
-              )
-              .padding(vertical = 4.dp, horizontal = 8.dp),
-            contentAlignment = Alignment.Center
-          ) {
-            Icon(
-              modifier = Modifier.size(12.dp),
-              painter = painterResource(R.drawable.ic_upvote),
-              tint = if (state.upvoted) {
-                HackerGreen
-              } else {
-                MaterialTheme.colorScheme.onSurface
-              },
-              contentDescription = "Upvote"
-            )
-          }
-        }
-        if (state.hidden == HiddenStatus.Displayed) {
-          Text(
-            text = state.content.parseAsHtml(),
-            style = MaterialTheme.typography.labelSmall,
-            fontWeight = FontWeight.Normal,
-            color = MaterialTheme.colorScheme.onSurface,
-          )
-        }
-      }
-    }
-
-    is CommentState.Loading -> {
-      val infiniteTransition = rememberInfiniteTransition("Skeleton")
-      val skeletonAlpha by infiniteTransition.animateFloat(
-        initialValue = 0.2f,
-        targetValue = 0.6f,
-        animationSpec = infiniteRepeatable(
-          animation = tween(durationMillis = 1000, easing = LinearEasing),
-          repeatMode = RepeatMode.Reverse,
-        ),
-        label = "Skeleton Alpha"
-      )
-      Column(
-        modifier = modifier
-          .fillMaxWidth()
-          .wrapContentHeight()
-          .clip(RoundedCornerShape(8.dp))
-          .background(color = MaterialTheme.colorScheme.surface)
-          .padding(8.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-      ) {
-        Row(
-          horizontalArrangement = Arrangement.spacedBy(8.dp),
-          verticalAlignment = Alignment.CenterVertically
-        ) {
-          Box(
-            modifier = Modifier
-              .width(40.dp)
-              .height(12.dp)
-              .clip(RoundedCornerShape(4.dp))
-              .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-          )
-
-          Row(horizontalArrangement = Arrangement.spacedBy(2.dp)) {
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-            )
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-            )
-          }
-          Spacer(modifier = Modifier.weight(1f))
-          Box(
-            modifier = Modifier
-              .clip(CircleShape)
-              .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.1f))
-              .padding(4.dp),
-            contentAlignment = Alignment.Center
-          ) {
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(CircleShape)
-                .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-            )
-          }
-        }
-        Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
-          Box(
-            modifier = Modifier
-              .fillMaxWidth()
-              .height(12.dp)
-              .clip(RoundedCornerShape(4.dp))
-              .background(MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-          )
-          Box(
-            modifier = Modifier
-              .fillMaxWidth()
-              .height(12.dp)
-              .clip(RoundedCornerShape(4.dp))
-              .background(MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-          )
-          Box(
-            modifier = Modifier
-              .fillMaxWidth(0.75f)
-              .height(12.dp)
-              .clip(RoundedCornerShape(4.dp))
-              .background(MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-          )
-        }
-      }
-    }
-  }
-}
-
-@PreviewLightDark
-@Composable
-fun CommentRowPreview() {
-  HackerNewsTheme {
-    Column {
-      CommentRow(
-        state = CommentState.Content(
-          id = 1,
-          level = 0,
-          author = "rikinm",
-          content = "Hello Parent",
-          timeLabel = "2d ago",
-          upvoted = false,
-          upvoteUrl = "",
-          children = listOf()
-        ),
-        onToggleHide = {},
-        onLikeTapped = {}
-      )
-    }
-  }
-}
-
-@PreviewLightDark
-@Composable
-fun CommentRowLoadingPreview() {
-  HackerNewsTheme {
-    Column {
-      CommentRow(
-        state = CommentState.Loading(level = 0),
-        onToggleHide = {},
-        onLikeTapped = {}
-      )
-    }
-  }
-}
-
-@Composable
-fun ItemHeader(
-  state: HeaderState,
-  modifier: Modifier = Modifier,
-  onLikeTapped: (HeaderState.Content) -> Unit,
-  onToggleBody: (Boolean) -> Unit
-) {
-  Column(
-    modifier = modifier.background(color = MaterialTheme.colorScheme.background),
-    verticalArrangement = Arrangement.spacedBy(8.dp)
-  ) {
-    when (state) {
-      is HeaderState.Content -> {
-        Column(verticalArrangement = Arrangement.spacedBy(16.dp)) {
-          Text(
-            text = state.title,
-            color = MaterialTheme.colorScheme.onSurface,
-            style = MaterialTheme.typography.titleSmall,
-            fontSize = 20.sp
-          )
-          Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-            verticalAlignment = Alignment.CenterVertically
-          ) {
-            Row(
-              horizontalArrangement = Arrangement.spacedBy(2.dp),
-              verticalAlignment = Alignment.CenterVertically
-            ) {
-              Text(
-                text = "@${state.author}",
-                color = HackerOrange,
-                style = MaterialTheme.typography.labelSmall,
-                fontWeight = FontWeight.Bold
-              )
-            }
-            MetadataTag(label = state.timeLabel) {
-              Icon(
-                modifier = Modifier.size(12.dp),
-                painter = painterResource(R.drawable.ic_time_outline),
-                tint = HackerPurple,
-                contentDescription = "Time Posted"
-              )
-            }
-            Spacer(modifier = Modifier.weight(1f))
-            MetadataButton(
-              label = "${state.points}",
-              contentColor = if (state.upvoted) {
-                HackerGreen
-              } else {
-                MaterialTheme.colorScheme.onSurface
-              },
-              onClick = { onLikeTapped(state) }
-            ) {
-              Icon(
-                modifier = Modifier.size(12.dp),
-                painter = painterResource(R.drawable.ic_upvote),
-                tint = HackerGreen,
-                contentDescription = "Upvotes"
-              )
-            }
-          }
-        }
-        if (state.body.text != null) {
-          Column(
-            Modifier
-              .fillMaxWidth()
-              .wrapContentHeight()
-              .clip(RoundedCornerShape(8.dp))
-              .clickable { onToggleBody(!state.body.collapsed) }
-              .background(color = MaterialTheme.colorScheme.surface)
-              .padding(8.dp),
-            verticalArrangement = Arrangement.spacedBy(4.dp)
-          ) {
-            Icon(
-              modifier = Modifier
-                .graphicsLayer {
-                  rotationZ = if (state.body.collapsed) 180f else 0f
-                }
-                .size(12.dp),
-              painter = painterResource(R.drawable.ic_collapse),
-              tint = MaterialTheme.colorScheme.onSurface,
-              contentDescription = "Expand or Collapse"
-            )
-            Text(
-              text = state.body.text.parseAsHtml(),
-              color = MaterialTheme.colorScheme.onBackground,
-              style = MaterialTheme.typography.labelSmall,
-              overflow = TextOverflow.Ellipsis,
-              maxLines = if (state.body.collapsed) 4 else Int.MAX_VALUE
-            )
-          }
-        }
-      }
-
-      HeaderState.Loading -> {
-        val infiniteTransition = rememberInfiniteTransition("Skeleton Loader")
-        val skeletonAlpha by infiniteTransition.animateFloat(
-          initialValue = 0.2f,
-          targetValue = 0.6f,
-          animationSpec = infiniteRepeatable(
-            animation = tween(durationMillis = 1000, easing = LinearEasing),
-            repeatMode = RepeatMode.Reverse
-          ),
-          label = "Skeleton Alpha"
-        )
-        Column(verticalArrangement = Arrangement.spacedBy(16.dp)) {
-          Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
-            Box(
-              modifier = Modifier
-                .fillMaxWidth()
-                .height(18.dp)
-                .clip(RoundedCornerShape(6.dp))
-                .background(color = MaterialTheme.colorScheme.onBackground.copy(alpha = skeletonAlpha))
-            )
-            Box(
-              modifier = Modifier
-                .fillMaxWidth(0.75f)
-                .height(18.dp)
-                .clip(RoundedCornerShape(6.dp))
-                .background(color = MaterialTheme.colorScheme.onBackground.copy(alpha = skeletonAlpha))
-            )
-          }
-        }
-        Row(
-          horizontalArrangement = Arrangement.spacedBy(4.dp),
-          verticalAlignment = Alignment.CenterVertically
-        ) {
-          Box(
-            modifier = Modifier
-              .width(60.dp)
-              .height(12.dp)
-              .clip(RoundedCornerShape(4.dp))
-              .background(HackerOrange.copy(alpha = skeletonAlpha))
-          )
-          Row(horizontalArrangement = Arrangement.spacedBy(2.dp)) {
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(HackerPurple.copy(alpha = skeletonAlpha))
-            )
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(MaterialTheme.colorScheme.onBackground.copy(alpha = skeletonAlpha))
-            )
-          }
-          Spacer(modifier = Modifier.weight(1f))
-          Row(
-            modifier = Modifier
-              .clip(CircleShape)
-              .background(MaterialTheme.colorScheme.onBackground.copy(alpha = 0.1f))
-              .padding(vertical = 4.dp, horizontal = 8.dp),
-            horizontalArrangement = Arrangement.spacedBy(2.dp)
-          ) {
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(HackerGreen.copy(alpha = skeletonAlpha))
-            )
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(MaterialTheme.colorScheme.onBackground.copy(alpha = skeletonAlpha))
-            )
-          }
-        }
-      }
-    }
-  }
-}
-
-@PreviewLightDark
-@Composable
-private fun ItemHeaderPreview() {
-  HackerNewsTheme {
-    Box(
-      modifier = Modifier
-        .background(MaterialTheme.colorScheme.background)
-        .padding(8.dp)
-    ) {
-      ItemHeader(
-        state = HeaderState.Content(
-          id = 0L,
-          title = "Show HN: A super neat HN client for Android",
-          author = "rikinm",
-          points = 69,
-          timeLabel = "2h ago",
-          body = BodyState("Wassup HN. I just built a sick new Hacker News Android client. Lalalalalala hahahah"),
-          upvoted = false,
-          upvoteUrl = "",
-        ),
-        modifier = Modifier
-          .fillMaxWidth()
-          .wrapContentHeight(),
-        onLikeTapped = {},
-        onToggleBody = {}
-      )
-    }
-  }
-}
-
-@PreviewLightDark
-@Composable
-private fun ItemHeaderLoadingPreview() {
-  HackerNewsTheme {
-    Box(
-      modifier = Modifier
-        .background(MaterialTheme.colorScheme.background)
-        .padding(8.dp)
-    ) {
-      ItemHeader(
-        state = HeaderState.Loading,
-        modifier = Modifier
-          .fillMaxWidth()
-          .wrapContentHeight(),
-        onLikeTapped = {},
-        onToggleBody = {}
-      )
-    }
-  }
-}
-
-@Composable
-fun PostCommentBump(
-  modifier: Modifier = Modifier,
-  state: PostCommentState,
-  goToLogin: () -> Unit,
-  updateComment: (String) -> Unit,
-  submitComment: () -> Unit
-) {
-  Column(
-    modifier = modifier
-      .clip(RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp))
-      .clickable {
-        if (!state.loggedIn) {
-          goToLogin()
-        }
-      }
-      .background(color = MaterialTheme.colorScheme.surfaceContainer)
-      .padding(vertical = 8.dp, horizontal = 32.dp),
-    verticalArrangement = Arrangement.spacedBy(8.dp, Alignment.CenterVertically),
-    horizontalAlignment = Alignment.CenterHorizontally
-  ) {
-    Row(
-      horizontalArrangement = Arrangement.spacedBy(8.dp),
-      verticalAlignment = Alignment.CenterVertically
-    ) {
-      Icon(
-        modifier = Modifier.size(12.dp),
-        painter = painterResource(R.drawable.ic_chat),
-        tint = MaterialTheme.colorScheme.onSurface,
-        contentDescription = "Add a comment"
-      )
-      Text(
-        text = "Add a comment",
-        style = MaterialTheme.typography.labelSmall,
-        color = MaterialTheme.colorScheme.onSurface
-      )
-    }
-    TextField(
-      modifier = Modifier.fillMaxWidth(),
-      value = state.text,
-      onValueChange = { updateComment(it) },
-      enabled = state.loggedIn,
-      keyboardOptions = KeyboardOptions(
-        keyboardType = KeyboardType.Text,
-        imeAction = ImeAction.Send
-      ),
-      keyboardActions = KeyboardActions(
-        onSend = { submitComment() }
-      ),
-      maxLines = 4,
-      shape = RoundedCornerShape(8.dp),
-      colors = TextFieldDefaults.colors(
-        focusedIndicatorColor = Color.Transparent,
-        unfocusedIndicatorColor = Color.Transparent,
-        disabledIndicatorColor = Color.Transparent,
-        errorIndicatorColor = Color.Transparent,
-        unfocusedContainerColor = MaterialTheme.colorScheme.surfaceContainerHigh,
-        focusedContainerColor = MaterialTheme.colorScheme.surfaceContainerHigh
-      ),
-    )
-  }
-}
