+++ android/app/src/main/java/com/emergetools/hackernews/features/stories/StoriesScreen.kt
@@ -1,49 +1,16 @@
 package com.emergetools.hackernews.features.stories
 
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.spring
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.combinedClickable
-import androidx.compose.foundation.horizontalScroll
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.Refresh
-import androidx.compose.material.icons.rounded.Warning
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.material3.pulltorefresh.PullToRefreshBox
 import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.runtime.Composable
@@ -53,35 +20,17 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.drawWithContent
-import androidx.compose.ui.draw.scale
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Path
 import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.PreviewLightDark
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.emergetools.hackernews.R
 import com.emergetools.hackernews.features.comments.CommentsDestinations
+import com.emergetools.hackernews.features.stories.components.FeedErrorCard
+import com.emergetools.hackernews.features.stories.components.FeedHeader
+import com.emergetools.hackernews.ui.components.ColumnSeparator
+import com.emergetools.hackernews.ui.components.StoryRow
 import com.emergetools.hackernews.ui.preview.AppStoreSnapshot
-import com.emergetools.hackernews.ui.theme.HackerBlue
-import com.emergetools.hackernews.ui.theme.HackerGreen
 import com.emergetools.hackernews.ui.theme.HackerNewsTheme
-import com.emergetools.hackernews.ui.theme.HackerOrange
-import com.emergetools.hackernews.ui.theme.HackerPurple
-import com.emergetools.hackernews.ui.theme.HackerRed
 import com.emergetools.snapshots.annotations.EmergeAppStoreSnapshot
-import me.saket.extendedspans.ExtendedSpans
-import me.saket.extendedspans.SquigglyUnderlineSpanPainter
-import me.saket.extendedspans.drawBehind
-import me.saket.extendedspans.rememberSquigglyUnderlineAnimator
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -170,7 +119,7 @@ fun StoriesScreen(
             },
           )
           if (index != state.stories.lastIndex) {
-            ListSeparator(
+            ColumnSeparator(
               lineColor = MaterialTheme.colorScheme.surfaceContainerHighest,
             )
           }
@@ -184,480 +133,34 @@ fun StoriesScreen(
 @Composable
 private fun StoriesScreenPreview() {
   HackerNewsTheme {
-    StoriesScreen(modifier = Modifier.fillMaxSize(), state = StoriesState(
-      stories = listOf(
-        StoryItem.Content(
-          id = 1L,
-          title = "Hello There",
-          author = "heyrikin",
-          score = 10,
-          commentCount = 0,
-          epochTimestamp = 100L,
-          timeLabel = "2h ago",
-          url = ""
-        ),
-        StoryItem.Content(
-          id = 1L,
-          title = "Hello There",
-          author = "heyrikin",
-          score = 10,
-          commentCount = 0,
-          epochTimestamp = 100L,
-          timeLabel = "2h ago",
-          url = ""
-        ),
-      )
-    ), actions = {}, navigation = {})
-  }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun StoryRow(
-  modifier: Modifier = Modifier,
-  item: StoryItem,
-  onClick: (StoryItem.Content) -> Unit,
-  onBookmark: (StoryItem.Content) -> Unit,
-  onCommentClicked: (StoryItem.Content) -> Unit,
-) {
-  when (item) {
-    is StoryItem.Content -> {
-      val bookmarkHeight by animateFloatAsState(
-        targetValue = if (item.bookmarked) {
-          80f
-        } else {
-          0f
-        }, animationSpec = spring(
-          dampingRatio = if (item.bookmarked) {
-            Spring.DampingRatioMediumBouncy
-          } else {
-            Spring.DampingRatioNoBouncy
-          }, stiffness = Spring.StiffnessLow
-        ), label = "Bookmark Height"
-      )
-      Column(
-        modifier = modifier
-          .fillMaxWidth()
-          .heightIn(min = 100.dp)
-          .background(color = MaterialTheme.colorScheme.background)
-          .drawWithContent {
-            drawContent()
-            val startX = size.width * 0.75f
-            val startY = 0f
-            val bookmarkWidth = 50f
-
-            val path = Path().apply {
-              moveTo(startX, startY)
-              lineTo(startX, startY + bookmarkHeight)
-              lineTo(startX + bookmarkWidth / 2f, startY + bookmarkHeight * 0.75f)
-              lineTo(startX + bookmarkWidth, startY + bookmarkHeight)
-              lineTo(startX + bookmarkWidth, startY)
-            }
-
-            drawPath(
-              path,
-              color = HackerOrange,
-            )
-          }
-          .combinedClickable(onClick = {
-            onClick(item)
-          }, onLongClick = {
-            onBookmark(item)
-          })
-          .padding(8.dp),
-        verticalArrangement = Arrangement.spacedBy(
-          space = 8.dp,
-          alignment = Alignment.CenterVertically
+    StoriesScreen(
+      modifier = Modifier.fillMaxSize(), state = StoriesState(
+        stories = listOf(
+          StoryItem.Content(
+            id = 1L,
+            title = "Hello There",
+            author = "heyrikin",
+            score = 10,
+            commentCount = 0,
+            epochTimestamp = 100L,
+            timeLabel = "2h ago",
+            url = ""
+          ),
+          StoryItem.Content(
+            id = 1L,
+            title = "Hello There",
+            author = "heyrikin",
+            score = 10,
+            commentCount = 0,
+            epochTimestamp = 100L,
+            timeLabel = "2h ago",
+            url = ""
+          ),
         )
-      ) {
-        Text(
-          text = "@${item.author}",
-          color = HackerOrange,
-          style = MaterialTheme.typography.labelSmall,
-          fontWeight = FontWeight.Bold
-        )
-        Text(
-          text = item.title,
-          color = MaterialTheme.colorScheme.onSurface,
-          style = MaterialTheme.typography.titleSmall
-        )
-        Spacer(modifier = Modifier.height(0.dp))
-        Row(
-          horizontalArrangement = Arrangement.spacedBy(8.dp),
-          verticalAlignment = Alignment.CenterVertically
-        ) {
-          MetadataTag(
-            label = "${item.score}",
-          ) {
-            Icon(
-              modifier = Modifier.size(12.dp),
-              painter = painterResource(R.drawable.ic_upvote),
-              tint = HackerGreen,
-              contentDescription = "Likes"
-            )
-          }
-          MetadataTag(
-            label = item.timeLabel,
-          ) {
-            Icon(
-              modifier = Modifier.size(12.dp),
-              painter = painterResource(R.drawable.ic_time_outline),
-              tint = HackerPurple,
-              contentDescription = "Time Posted"
-            )
-          }
-          Spacer(modifier = Modifier.weight(1f))
-          MetadataButton(
-            label = "${item.commentCount}",
-            onClick = { onCommentClicked(item) }
-          ) {
-            Icon(
-              modifier = Modifier.size(12.dp),
-              painter = painterResource(R.drawable.ic_chat),
-              tint = HackerBlue,
-              contentDescription = "Comments"
-            )
-          }
-        }
-      }
-    }
-
-    is StoryItem.Loading -> {
-      val infiniteTransition = rememberInfiniteTransition("Skeleton")
-      val skeletonAlpha by infiniteTransition.animateFloat(
-        initialValue = 0.2f,
-        targetValue = 0.6f,
-        animationSpec = infiniteRepeatable(
-          animation = tween(durationMillis = 1000, easing = LinearEasing),
-          repeatMode = RepeatMode.Reverse,
-        ),
-        label = "Skeleton Alpha"
-      )
-      Column(
-        modifier = modifier
-          .fillMaxWidth()
-          .heightIn(min = 100.dp)
-          .background(color = MaterialTheme.colorScheme.background)
-          .padding(8.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp, Alignment.CenterVertically)
-      ) {
-        Box(
-          modifier = Modifier
-            .fillMaxWidth(0.15f)
-            .height(12.dp)
-            .clip(RoundedCornerShape(4.dp))
-            .background(color = HackerOrange.copy(alpha = skeletonAlpha))
-        )
-        Column(verticalArrangement = Arrangement.spacedBy(2.dp)) {
-          Box(
-            modifier = Modifier
-              .fillMaxWidth(0.75f)
-              .height(12.dp)
-              .clip(RoundedCornerShape(4.dp))
-              .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-          )
-          Box(
-            modifier = Modifier
-              .fillMaxWidth(0.5f)
-              .height(12.dp)
-              .clip(RoundedCornerShape(4.dp))
-              .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-          )
-        }
-        Spacer(modifier = Modifier.height(0.dp))
-        Row(
-          horizontalArrangement = Arrangement.spacedBy(8.dp),
-          verticalAlignment = Alignment.CenterVertically
-        ) {
-          Row(
-            modifier = Modifier.wrapContentSize(),
-            horizontalArrangement = Arrangement.spacedBy(2.dp)
-          ) {
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(color = HackerGreen.copy(alpha = skeletonAlpha))
-            )
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-            )
-          }
-          Row(
-            modifier = Modifier.wrapContentSize(),
-            horizontalArrangement = Arrangement.spacedBy(2.dp)
-          ) {
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(color = HackerPurple.copy(alpha = skeletonAlpha))
-            )
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-            )
-          }
-          Spacer(modifier = Modifier.weight(1f))
-          Row(
-            modifier = Modifier
-              .wrapContentSize()
-              .clip(CircleShape)
-              .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.1f))
-              .padding(vertical = 4.dp, horizontal = 8.dp),
-            horizontalArrangement = Arrangement.spacedBy(2.dp)
-          ) {
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(color = HackerBlue.copy(alpha = skeletonAlpha))
-            )
-            Box(
-              modifier = Modifier
-                .size(12.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(color = MaterialTheme.colorScheme.onSurface.copy(alpha = skeletonAlpha))
-            )
-          }
-        }
-      }
-    }
-  }
-}
-
-@Composable
-fun ListSeparator(
-  lineColor: Color,
-  space: Dp = 0.5.dp
-) {
-  Spacer(
-    modifier = Modifier
-      .fillMaxWidth()
-      .height(space)
-      .background(color = lineColor)
-  )
-}
-
-@Composable
-fun MetadataButton(
-  label: String,
-  contentColor: Color = MaterialTheme.colorScheme.onSurface,
-  backgroundColor: Color = contentColor.copy(alpha = 0.1f),
-  onClick: () -> Unit = {},
-  icon: @Composable () -> Unit,
-) {
-  Row(
-    modifier = Modifier
-      .wrapContentSize()
-      .clip(CircleShape)
-      .clickable { onClick() }
-      .background(color = backgroundColor)
-      .padding(vertical = 4.dp, horizontal = 8.dp),
-    horizontalArrangement = Arrangement.spacedBy(4.dp, Alignment.CenterHorizontally),
-    verticalAlignment = Alignment.CenterVertically
-  ) {
-    icon()
-    Text(
-      text = label,
-      style = MaterialTheme.typography.labelSmall,
-      fontWeight = FontWeight.Medium,
-      color = contentColor
-    )
-  }
-}
-
-@Composable
-fun MetadataTag(
-  label: String,
-  contentColor: Color = MaterialTheme.colorScheme.onSurface,
-  icon: @Composable () -> Unit
-) {
-  Row(
-    modifier = Modifier.wrapContentSize(),
-    horizontalArrangement = Arrangement.spacedBy(2.dp, Alignment.CenterHorizontally),
-    verticalAlignment = Alignment.CenterVertically
-  ) {
-    icon()
-    Text(
-      text = label,
-      style = MaterialTheme.typography.labelSmall,
-      fontWeight = FontWeight.Medium,
-      color = contentColor
-    )
-  }
-}
-
-@PreviewLightDark
-@Composable
-private fun StoryRowPreview() {
-  HackerNewsTheme {
-    StoryRow(
-      item = StoryItem.Content(
-        id = 1L,
-        title = "Hello There",
-        author = "heyrikin",
-        score = 10,
-        commentCount = 0,
-        epochTimestamp = 100L,
-        timeLabel = "2h ago",
-        bookmarked = true,
-        url = ""
       ),
-      onClick = {},
-      onBookmark = {},
-      onCommentClicked = {},
-    )
+      actions = {},
+      navigation = {})
   }
-}
-
-@PreviewLightDark
-@Composable
-private fun StoryRowLoadingPreview() {
-  HackerNewsTheme {
-    StoryRow(
-      item = StoryItem.Loading(id = 1L),
-      onClick = {},
-      onBookmark = {},
-      onCommentClicked = {},
-    )
-  }
-}
-
-@Composable
-fun FeedErrorCard(modifier: Modifier = Modifier, onRefresh: () -> Unit) {
-  Column(
-    modifier = modifier
-      .fillMaxWidth()
-      .height(200.dp)
-      .background(color = MaterialTheme.colorScheme.background),
-    verticalArrangement = Arrangement.spacedBy(8.dp, Alignment.CenterVertically),
-    horizontalAlignment = Alignment.CenterHorizontally
-  ) {
-    Icon(
-      imageVector = Icons.Rounded.Warning,
-      tint = HackerRed,
-      contentDescription = "Failed to Load",
-    )
-
-    Text(
-      text = "Failed to Load Feed",
-      style = MaterialTheme.typography.titleSmall,
-      color = MaterialTheme.colorScheme.onBackground
-    )
-
-    Button(
-      colors = ButtonDefaults.buttonColors(
-        containerColor = HackerRed,
-        contentColor = Color.White
-      ),
-      onClick = { onRefresh() }) {
-      Icon(
-        imageVector = Icons.Rounded.Refresh,
-        contentDescription = "Reload Feed"
-      )
-    }
-  }
-}
-
-@PreviewLightDark
-@Composable
-fun FeedErrorCardPreview() {
-  HackerNewsTheme {
-    FeedErrorCard(
-      onRefresh = {}
-    )
-  }
-}
-
-@Composable
-fun FeedHeader(
-  modifier: Modifier = Modifier,
-  feeds: List<NewsFeed>,
-  onSelected: (FeedType) -> Unit
-) {
-  val scrollState = rememberScrollState()
-
-  Row(
-    modifier = modifier
-      .horizontalScroll(scrollState)
-      .fillMaxWidth()
-      .wrapContentHeight()
-      .background(color = MaterialTheme.colorScheme.background),
-    horizontalArrangement = Arrangement.spacedBy(4.dp, Alignment.CenterHorizontally),
-    verticalAlignment = Alignment.CenterVertically
-  ) {
-    feeds.forEach { feed ->
-      FeedHeaderItem(state = feed, select = onSelected)
-    }
-  }
-}
-
-@PreviewLightDark
-@Composable
-fun FeedHeaderPreview() {
-  HackerNewsTheme {
-    FeedHeader(
-      feeds = supportedFeeds,
-      onSelected = {}
-    )
-  }
-}
-
-@Composable
-fun FeedHeaderItem(state: NewsFeed, select: (FeedType) -> Unit) {
-  val squigglyUnderlineAnimator = rememberSquigglyUnderlineAnimator()
-  val spans = remember(state.selected) {
-    ExtendedSpans(
-      SquigglyUnderlineSpanPainter(
-        width = if (state.selected) 2.sp else 0.sp,
-        amplitude = if (state.selected) 1.sp else 0.sp,
-        animator = squigglyUnderlineAnimator
-      )
-    )
-  }
-  val label = remember(state) {
-    AnnotatedString(
-      text = state.type.label,
-      spanStyle = if (state.selected) {
-        SpanStyle(textDecoration = TextDecoration.Underline)
-      } else {
-        SpanStyle()
-      }
-    )
-  }
-
-  val scale by animateFloatAsState(
-    targetValue = if (state.selected) 1.0f else 0.8f,
-    animationSpec = spring(
-      dampingRatio = Spring.DampingRatioMediumBouncy,
-      stiffness = Spring.StiffnessLow
-    ),
-    label = "Selection Scale"
-  )
-  Text(
-    modifier = Modifier
-      .scale(scale)
-      .clickable(
-        interactionSource = remember { MutableInteractionSource() },
-        indication = null
-      ) {
-        select(state.type)
-      }
-      .padding(8.dp)
-      .drawBehind(spans),
-    text = spans.extend(label),
-    color = if (state.selected) HackerOrange else MaterialTheme.colorScheme.onBackground.copy(alpha = 0.2f),
-    style = MaterialTheme.typography.titleMedium,
-    onTextLayout = { result ->
-      spans.onTextLayout(result)
-    }
-  )
 }
 
 @OptIn(EmergeAppStoreSnapshot::class)
@@ -768,8 +271,13 @@ private fun StoriesScreenAppStorePreview() {
   )
 
   HackerNewsTheme {
-    StoriesScreen(modifier = Modifier.fillMaxSize(), state = StoriesState(
-      stories = stories
-    ), actions = {}, navigation = {})
+    StoriesScreen(
+      modifier = Modifier.fillMaxSize(),
+      state = StoriesState(
+        stories = stories
+      ),
+      actions = {},
+      navigation = {}
+    )
   }
 }
